requirements_fixed.txt
openai-whisper==20240930
torch==2.0.0
torchaudio==2.0.0
pyperclip==1.8.2
pathlib